Updated to new snake case naming convention Quantconnect uses, code now works agian.
Furthermore, fixed a bunch of bugs/inconsistencies with the logic/performance.

Default values not use a trend following strategy, with no custom tweaks, so you can start out on a solid foundation.

For the customized strategy, see the previous broken commit and its variables.

Switched resoltuion from hourly to daily.